chore: remove europa satellite (front-end microapp)
