+++ gulpfile.js
@@ -7,7 +7,7 @@ var babelify   = require('babelify')
 var source     = require('vinyl-source-stream')
 var buffer     = require('vinyl-buffer')
 
-var satellites = ['amalthea', 'callisto', 'europa', 'ganymede', 'io']
+var satellites = ['amalthea', 'callisto', 'ganymede', 'io']
 var environment = process.env.NODE_ENV
 
 gulp.task('styles', function () {
@@ -17,7 +17,6 @@ gulp.task('styles', function () {
             './web/core/styles/errors.scss',
             './web/amalthea/styles/amalthea.scss',
             './web/callisto/styles/callisto.scss',
-            './web/europa/styles/europa.scss',
             './web/ganymede/styles/ganymede.scss',
             './web/io/styles/io.scss',
         ])
