--- web/europa/styles/europa.scss
@@ -1,2 +1,0 @@
-@import "jupiter";
-@import "settings";
