--- web/europa/styles/settings.scss
@@ -1,218 +1,0 @@
-.settings-wrapper {
-  margin-bottom: 40px;
-}
-.side-nav {
-  .divider {
-    margin-bottom: 1rem;
-  }
-}
-.application-card {
-  padding: 15px 15px;
-  border: 1px solid rgba(0, 0, 0, 0.5);
-
-  .title {
-    font-size: 1.3rem;
-    line-height: 1;
-  }
-
-  .scope {
-    font-size: 0.8rem;
-    font-style: italic;
-  }
-
-  .action {
-    float: right;
-    margin-bottom: 0;
-  }
-
-  p {
-    margin-bottom: 0.25rem;
-  }
-}
-.application-card ~ .application-card {
-  margin-top: 12px;
-}
-.profile {
-  @media screen and (min-width: 600px) {
-    box-shadow: 0px 2px 2px rgba(225,225,225,0.2);
-  }
-  background: #f7f7f7;
-  color: #222;
-
-  .profile-entry {
-    padding: 0.9rem 0 0.7625rem;
-
-    .field {
-      font-weight: 500;
-    }
-
-    .value {
-      font-weight: 200;
-    }
-  }
-
-  .profile-entry ~ .profile-entry {
-    border-top: 1px solid rgba(0,0,0,0.12);
-  }
-}
-.profile-actions {
-  padding: 18px 0;
-}
-.token-error {
-  text-align: center;
-  padding: 20px 0;
-}
-.thin-input {
-  max-width: 350px;
-  height: auto;
-  margin-bottom: 0.5rem;
-  padding: 0.3rem 0.5rem;
-}
-table {
-  width: 100%;
-  border: none;
-  background: transparent;
-  color: #878787;
-
-  thead {
-    background: transparent;
-
-    tr th {
-      padding: 1.4rem 1.2rem;
-      border-top: 1px solid #d7d7d7;
-      border-bottom: 1px solid #d7d7d7;
-      font-weight: 500;
-      text-transform: uppercase;
-      color: inherit;
-    }
-  }
-
-  tr.odd, tr:nth-of-type(odd) {
-    td {
-      color: inherit;
-    }
-
-    color: inherit;
-  }
-}
-.blank-list {
-  text-align: center;
-}
-.jupiter-accordion {
-  .jupiter-accordion-title {
-    position: relative;
-    font-size: 0.9rem;
-    font-weight: normal;
-    border-bottom: 1px solid #e6e6e6;
-    margin: 0;
-
-    a {
-      display: block;
-      color: inherit;
-      padding: 6px 8px;
-      padding-left: 26px;
-
-      &::before {
-        content: "\f107";
-        position: absolute;
-        left: 6px;
-        font-family: "Font Awesome 5 Free";
-        font-style: normal;
-        font-weight: 900;
-      }
-    }
-  }
-
-  .jupiter-accordion-body {
-    max-height: 0;
-    overflow: hidden;
-    transition: 0.5s ease;
-  }
-
-  .jupiter-accordion-child.open {
-    .jupiter-accordion-title {
-      a::before {
-        content: "\f106";
-      }
-    }
-
-    .jupiter-accordion-body {
-      padding: 16px 12px;
-      max-height: initial;
-      transition: 0.5s ease;
-    }
-  }
-}
-.new-application {
-  .title {
-    font-weight: 300;
-    font-size: 1.75rem;
-    text-align: right;
-  }
-
-  .description {
-    text-align: right;
-  }
-
-  [type='text'],
-  [type='url'],
-  textarea {
-    height: auto;
-    padding: 0.25rem 0.5rem;
-    margin: 0 0 0.75rem;
-    resize: none;
-  }
-}
-.applications-listing {
-  margin-bottom: 20px;
-
-  .inline-list {
-    margin: 8px 0 0;
-    font-size: 0.9rem;
-
-    > li:first-of-type {
-      padding-left: 0;
-    }
-  }
-}
-.applications-divisor {
-  .applications-divisor-title {
-    color: #2199e8;
-    position: relative;
-    font-size: 0.9rem;
-    font-weight: normal;
-    border-bottom: 1px solid #2199e8;
-    margin: 0 0 26px;
-
-    span {
-      display: block;
-      color: inherit;
-      padding: 6px 8px;
-      padding-left: 26px;
-
-      &::before {
-        content: "\f105";
-        position: absolute;
-        left: 7px;
-        font-family: "Font Awesome 5 Free";
-        font-style: normal;
-        font-weight: 900;
-      }
-    }
-  }
-}
-.form-common.internal {
-  padding: 12px 0;
-
-  .button-wrapper {
-    margin-top: 20px;
-  }
-
-  .button-anchor {
-    margin: 10px 0;
-  }
-
-  .form-warning {
-    font-size: 0.9rem;
-  }
-}
